+++ pages/integrations/xdd/extractions/@paperId/+Page.ts
@@ -3,12 +3,15 @@ import h from "@macrostrat/hyper";
 import { ContentPage } from "~/layouts";
 import { PageBreadcrumbs } from "~/components";
 import { usePageContext } from "vike-react/usePageContext";
-import { ExtractionContext, enhanceData } from "../lib";
+import {
+  ExtractionContext,
+  enhanceData,
+} from "@macrostrat/feedback-components/src/extractions";
 import {
   useEntityTypeIndex,
   useModelIndex,
   usePostgresQuery,
-} from "../lib/data-service";
+} from "./data-service";
 
 export function Page() {
   return h(ContentPage, [h(PageBreadcrumbs), h(PageMain)]);
