--- pages/integrations/xdd/extractions/lib/main.module.sass
@@ -1,10 +1,0 @@
-.entities
-  list-style: none
-  padding-left: 0
-
-  .entities ul
-    list-style: none
-
-.entity
-  margin: 0.2em 0 0.5em
-  padding-right: 3px
