+++ pages/integrations/xdd/feedback/@sourceTextID/+Page.client.module.sass
@@ -1,0 +1,13 @@
+.feedback-main
+  display: flex
+  flex-direction: column
+  max-width: 70em
+  margin: 0 auto
+  position: relative
+  max-height: 100%
+
+.feedback-interface
+  display: flex
+  flex-direction: column
+  flex: 1
+  min-height: 100px
