--- pages/integrations/xdd/feedback/@sourceTextID/lib/feedback.module.sass
@@ -1,52 +1,0 @@
-.feedback-component
-  position: relative
-  width: 800px
-
-  & > svg
-    width: 800px
-
-.node
-  cursor: pointer
-
-
-.feedback-text
-  margin-bottom: 2em
-
-.entity-panel
-  position: relative
-
-.control-panel
-  max-width: 15em
-  position: absolute
-  top: 1em
-  right: 1em
-  padding: 0.2em 0.5em
-
-
-.feedback-main
-  display: flex
-  flex-direction: column
-  max-width: 70em
-  margin: 0 auto
-  position: relative
-  max-height: 100%
-
-.feedback-interface
-  display: flex
-  flex-direction: column
-  flex: 1
-  min-height: 100px
-
-
-.entity-panel
-  flex: 1
-  min-height: 100px
-  padding: 1em
-  background: var(--panel-secondary-background-color)
-  border-radius: 4px
-  // Inset box shadow
-  box-shadow: 0 0 0 1px var(--panel-border-color) inset
-
-.selection-tree
-  margin: -1em 0
-  padding: 1em 0
