--- pages/integrations/xdd/feedback/@sourceTextID/lib/type-selector/main.module.sass
@@ -1,13 +1,0 @@
-.item-container
-  min-height: 40px
-  display: flex
-  align-items: center
-  gap: 1em
-
-  &.selected
-    background-color: var(--panel-background-color)
-
-.swatch
-  width: 30px
-  height: 30px
-  border-radius: 4px
